pyproject.toml
[project]
name = "nonebot-plugin-laofei-tools"
version = "0.1.0"
description = "老肥工具箱 - NoneBot 插件，支持以图搜图等功能"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your@email.com" }
]
keywords = ["nonebot", "nonebot2", "plugin", "image-search", "soutubot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "nonebot2>=2.0.0",
    "nonebot-adapter-onebot>=2.0.0",
    "httpx>=0.24.0",
    "pillow>=10.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
]

[project.entry-points."nonebot.plugins"]
nonebot_plugin_laofei_tools = "nonebot_plugin_laofei_tools:__init__"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nonebot_plugin_laofei_tools"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
